\name{AmpData}
\docType{instance}
\alias{AmpData}
\title{AffyBatch instance with six arrays from amplication experiment}
\description{
This \code{\link{AffyBatch}} represents part of an expeirment to
determine if one can use amplified RNA with Affymetrix GeneChip arrays.
}
\usage{data(AmpData)}
\format{A \code{\link{AffyBatch}} containing 6 arrays.}
\source{Anonymous for now...}
\keyword{datasets}
