\name{AmpData}
\docType{data}
\alias{AmpData}
\title{AffyBatch instance with six arrays from amplication experiment}
\description{
This \code{\link[affy:AffyBatch-class]{AffyBatch}} represents part of an
expeirment to determine if one can use amplified RNA with Affymetrix
GeneChip arrays. 
}
\usage{data(AmpData)}
\format{A \code{\link[affy:AffyBatch-class]{AffyBatch}} containing 6 arrays.}
\source{Anonymous for now...}
\keyword{datasets}
